fix(scanner): split audit-log reader out of http.js
ClawHub's static-scanner v2.4.2 flagged
suspicious.potential_exfiltration on lib/http.js:105 because
readOutboundLog (file read) co-located with the fetch-based network
sends in the same file. The scanner heuristic is "readFile + network
send in same module = potential exfiltration", regardless of whether
the file being read is our own audit log.

Fix: move readOutboundLog (and the audit-log path constants) to
a new lib/audit.js. http.js now only has the write side
(logOutbound: appendFileSync + statSync + renameSync + mkdirSync
+ existsSync — no readFileSync). privacy.js imports
readOutboundLog from lib/audit.js.

No functional change. /mapick privacy log behaves identically.
